Move domain objects to different package
